Split operator for readability
The code of the operator was in a massive single file.

To improve readability, this was split into multiple files.

It also took the liberty to merge the anonymous function with the
handler doing approximatively the same work, as none of those
were tested.
